cleanup g3_* etc

+++ lib/gnu_gama/adj/adj.cpp
@@ -91,7 +91,7 @@ void AdjInputData::write_xml(std::ostream& out) const
 
   if (pcov)
     {
-      const long blocks = pcov->blocks();
+      const size_t blocks = pcov->blocks();
 
       out << "\n  <block-diagonal>\n"
           << "    <blocks>" << blocks << "</blocks>"
@@ -99,8 +99,8 @@ void AdjInputData::write_xml(std::ostream& out) const
 
       for (long b=1; b<=blocks; b++)
         {
-          long dim   = pcov->dim(b);
-          long width = pcov->width(b);
+          size_t dim   = pcov->dim(b);
+          size_t width = pcov->width(b);
 
           out << "      <block> <dim>"
               << dim    << "</dim> <width>"
@@ -160,7 +160,7 @@ void AdjInputData::read_xml(std::istream& inp)
   while (getline(inp, line))
     {
       line += '\n';
-      dp.xml_parse(line.c_str(), line.length(), 0);
+      dp.xml_parse(line.c_str(), int(line.length()), 0);
     }
   dp.xml_parse("", 0, 1);
 
@@ -230,8 +230,8 @@ void AdjInputData::read_gama_local_old_format(std::istream& inp)
   for (long k=0, r=1; r<=rows; r++)
     {
       A->new_row();
-      long nonz = *m++;
-      for (long i=1; i<=nonz; i++, k++)  A->add_element(flt[k], ind[k]);
+      size_t nonz = *m++;
+      for (size_t i=1; i<=nonz; i++, k++)  A->add_element(flt[k], ind[k]);
     }
 
   delete pcov;
