+++ lib/gnu_gama/g3/g3_model.h
@@ -179,7 +179,7 @@ namespace GNU_gama {
     // adjustment
     Adj*              adj;
 
-    int    redundancy;
+    size_t redundancy;
     enum { apriori, aposteriori } actual_sd;
     double aposteriori_sd;
     double std_deviation;
